model_server/florence2/requirements.txt
torch>=2.1.0
transformers>=4.40.0
fastapi>=0.100.0
uvicorn>=0.19.0
httpx>=0.27.0
pillow>=9.3.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
